% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_KITra.R
\name{KITra_palettes}
\alias{KITra_palettes}
\title{KITra_palettes}
\usage{
KITra_palettes(
  name,
  n,
  all_palettes = KITra_colours,
  type = c("discrete", "continuous")
)
}
\arguments{
\item{name}{The name of the palette to extract.}

\item{n}{The number of colors to extract from the palette. If not specified, the entire palette is used.}

\item{all_palettes}{The list of palettes to extract the palette from. Default is `KITra_colours`.}

\item{type}{The type of palette to return. Can be "discrete" or "continuous". Default is "discrete".}
}
\value{
A palette object, which can be used as an argument in ggplot2's `scale_color_manual` or `scale_fill_manual` functions.
}
\description{
Extracts a specific palette from the `KITra_colours` list and applies it as a color or fill scale in a ggplot2 plot.
}
