% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_KITra.R
\name{scale_fill_KITra_continuous}
\alias{scale_fill_KITra_continuous}
\title{scale_fill_KITra_continuous}
\usage{
scale_fill_KITra_continuous(name = "gradient")
}
\arguments{
\item{name}{The name of the palette to use. Default is "gradient".}
}
\value{
A scale_fill_manual function with the specified palette.
}
\description{
A convenience function that applies a specific continuous palette to the fill scale of a ggplot2 plot.
}
