% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_KITra.R
\name{scale_fill_KITra_discrete}
\alias{scale_fill_KITra_discrete}
\title{scale_fill_KITra_discrete}
\usage{
scale_fill_KITra_discrete(name = "cat")
}
\arguments{
\item{name}{The name of the palette to use. Default is "cat".}
}
\value{
A scale_fill_manual function with the specified palette.
}
\description{
A convenience function that applies a specific discrete palette to the fill scale of a ggplot2 plot.
}
