% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_KITra.R
\name{theme_KITra}
\alias{theme_KITra}
\title{theme_KITra}
\usage{
theme_KITra(base_size = 12)
}
\arguments{
\item{base_size}{The base text size to use in the theme. Default is 12.}
}
\value{
A ggplot2 theme object that can be added to a plot using the `+` operator.
}
\description{
Applies a custom theme to a ggplot2 plot, with custom colors for text and grid elements.
}
